+++ db/DbOfficialReplaceManager.php
@@ -449,7 +449,7 @@ class DbOfficialReplaceManager {
         $basePath = dirname($requestUri);
         $basePath = $basePath === '/' ? '' : $basePath;
         $selfUrl = $scheme . '://' . $host . $basePath;
-        return $selfUrl . '/mx.php?action=mxjx&url=' . urlencode($m3u8Url);
+        return $selfUrl . '/mx.php?action=mxjx&deep=1&url=' . urlencode($m3u8Url);
     }
 
     private function logResolve($url, $platform, $title, $score, $site, $m3u8Url, $success) {